커스텀 오버레이 및 컴포넌트 분리
